+++ src/algo.h
@@ -1,16 +1,11 @@
 #ifndef ALGO_H_   /* Include guard */
 #define ALGO_H_
 
-int algorithme1(int entier);
-
-int algorithme2(int entier);
-
-int algorithme3(int entier);
-
-int algorithme4(int entier);
-
-int algorithme5(int entier);
-
-int algorithme6(int entier);
+long int  algorithme1(long int entier);
+long int  algorithme2(long int entier);
+long int  algorithme3(long int entier);
+long int  algorithme4(long int entier);
+long int  algorithme5(long int entier);
+long int  algorithme6(long int entier);
 
 #endif
